requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
playwright==1.40.0
apscheduler==3.10.4
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
bcrypt==4.1.2
